requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aiofiles==0.7.0
    # via quart
asyncpg==0.24.0
    # via -r requirements.in
blinker==1.4
    # via quart
click==8.0.1
    # via quart
greenlet==1.1.1
    # via sqlalchemy
h11==0.12.0
    # via
    #   hypercorn
    #   wsproto
h2==4.0.0
    # via hypercorn
hpack==4.0.0
    # via h2
hypercorn==0.11.2
    # via quart
hyperframe==6.0.1
    # via h2
itsdangerous==2.0.1
    # via quart
jinja2==3.0.1
    # via quart
markupsafe==2.0.1
    # via jinja2
priority==2.0.0
    # via hypercorn
python-environ==0.4.54
    # via -r requirements.in
quart==0.15.1
    # via -r requirements.in
sqlalchemy==1.4.23
    # via -r requirements.in
toml==0.10.2
    # via
    #   hypercorn
    #   quart
werkzeug==2.0.1
    # via quart
wsproto==1.0.0
    # via hypercorn
